requirements.txt
# Financial Report Analyzer - Dependencies

# SEC EDGAR
sec-edgar-downloader>=5.0.0

# Vector Database
chromadb>=0.4.0

# Document Processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16
lxml>=4.9.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Clients
requests>=2.31.0
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# GUI
streamlit>=1.30.0
